requirements.txt
truthbrush
pandas
psycopg2-binary
sqlalchemy
seaborn
tqdm
pika
dash
black
fastapi
uvicorn
python-jose
flask
python-multipart
dash-cytoscape
networkx
python-louvain
beautifulsoup4
python-dotenv